+++ app/src/main/java/com/godzuche/achivitapp/feature_task/presentation/ui_elements/modal_bottom_sheet/ModalBottomSheetUiEvent.kt
@@ -1,5 +1,5 @@
 package com.godzuche.achivitapp.feature_task.presentation.ui_elements.modal_bottom_sheet
 
 sealed class ModalBottomSheetUiEvent {
-    data class OnGetBottomSheetAction(val taskId: Long) : ModalBottomSheetUiEvent()
+    data class OnGetBottomSheetAction(val taskId: Int) : ModalBottomSheetUiEvent()
 }
