+++ app/src/main/java/com/godzuche/achivitapp/feature_task/presentation/ui_elements/modal_bottom_sheet/ModalBottomSheetUiState.kt
@@ -5,6 +5,6 @@ import com.godzuche.achivitapp.feature_task.domain.model.Task
 data class ModalBottomSheetUiState(
     val bottomSheetAction: String = "",
     val task: Task? = null,
-    val id: Long = -1L,
+    val id: Int = -1,
 )
